ccg015 focus item toppage after remove 1 item

+++ nts.uk/web/nts.uk.com.web/src/main/webapp/view/ccg/015/a/ccg015.a.vm.js
@@ -49,13 +49,18 @@ var nts;
                                     var dfd = $.Deferred();
                                     self.listTopPage([]);
                                     a.service.loadTopPage().done(function (data) {
-                                        data.forEach(function (item, index) {
-                                            self.listTopPage.push(new Node(item.topPageCode, item.topPageName, null));
-                                            dfd.resolve();
-                                        });
-                                        if (self.listTopPage().length > 0) {
-                                            self.toppageSelectedCode(self.listTopPage()[0].code);
+                                        if (data.length > 0) {
+                                            data.forEach(function (item, index) {
+                                                self.listTopPage.push(new Node(item.topPageCode, item.topPageName, null));
+                                            });
+                                            if (self.listTopPage().length > 0) {
+                                                self.toppageSelectedCode(self.listTopPage()[0].code);
+                                            }
                                         }
+                                        else {
+                                            self.newTopPage();
+                                        }
+                                        dfd.resolve();
                                     });
                                     return dfd.promise();
                                 };
@@ -103,6 +108,8 @@ var nts;
                                     }).onClosed(function () {
                                     });
                                 };
+                                ScreenModel.prototype.openFlowMenuSetting = function () {
+                                };
                                 ScreenModel.prototype.newTopPage = function () {
                                     var self = this;
                                     self.topPageModel(new TopPageModel());
@@ -110,9 +117,33 @@ var nts;
                                 };
                                 ScreenModel.prototype.removeTopPage = function () {
                                     var self = this;
-                                    a.service.deleteTopPage(self.toppageSelectedCode()).done(function () {
-                                    }).fail();
-                                    self.loadTopPageList();
+                                    nts.uk.ui.dialog.confirm("//TODO #Msg_18").ifYes(function () {
+                                        var removeCode = self.toppageSelectedCode();
+                                        var removeIndex = self.getIndexOfRemoveItem(removeCode);
+                                        var listLength = self.listTopPage().length;
+                                        a.service.deleteTopPage(self.toppageSelectedCode()).done(function () {
+                                        }).fail();
+                                        self.loadTopPageList().done(function () {
+                                            var lst = self.listTopPage();
+                                            if (removeIndex < listLength - 1) {
+                                                self.toppageSelectedCode(lst[removeIndex].code);
+                                            }
+                                            else {
+                                                self.toppageSelectedCode(lst[removeIndex - 1].code);
+                                            }
+                                        });
+                                    }).ifNo(function () {
+                                    });
+                                };
+                                ScreenModel.prototype.getIndexOfRemoveItem = function (code) {
+                                    var self = this;
+                                    var ind = 0;
+                                    self.listTopPage().forEach(function (item, index) {
+                                        if (item.code == code) {
+                                            ind = index;
+                                        }
+                                    });
+                                    return ind;
                                 };
                                 return ScreenModel;
                             }());
